Added a simple JSON api that allows clients to search for trustees by twitter username.
Example URL:

  /users/search?twitter_username=toolness

On my test database, this query returns:

  [{"first_name":"Poseidon",
    "last_name":"Johnson",
    "twitter_username":"toolness",
    "id":1}]

If no users were matched, an empty list would be returned.

This API can be extended later to search for users by other parameters.

Thanks to Jesse Farmer and Raph for helping me with most of this, as I don't know Ruby.